New method for working out stable cycles
New threshold used on dVdt data when determining the stable cycles. Values are filtered using a moving average, and then peaks are found for which prominence is greater than whichever is larger: some % of the maximum dVdt, or some multiple of the standard deviation of the gas signal.
The user will now be alerted when no peaks are found, which is assumed to occur when the laser doesn't fire.